+++ PSGSuite/Aliases/PSGSuite.Aliases.ps1
@@ -42,5 +42,4 @@
     'Update-GSGmailSendAsSettings'      = 'Update-GSGmailSendAsAlias'
     'Update-GSSheetValue'               = 'Export-GSSheet'
     'Update-GSTeamDrive'                = 'Update-GSDrive'
-    'Get-PSGSuiteScope'                 = 'Get-PSGSuiteOAuthScope'
 }